+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-# 0.0.1
+
+## 0.0.3
+
+* fix convert datetime
+* add check location activate
+* add open view activate location
+
+## 0.0.2
+
+* Windows.Devices.Geolocation: Latitude and Longitude;
+
+## 0.0.1
 
 * TODO: Internationalization for Windows Applications.
 * SYSGEOCLASS: Specifies the geographical location class.
 * SYSGEOTYPE: Defines the type of geographical location information requested in the GetGeoInfo or GetGeoInfoEx function.
 
-## 0.0.2
 
-* Windows.Devices.Geolocation: Latitude and Longitude;
+